Replace big search algorithm with simple text scoring algorithm
Deletes the multiple functions that calculated a score for pages.
Instead uses a simpler function, that calculates a score for text in general,
which now gets used on title and text of pages.
The new score is based on number of search words in text, and their length.
Deletes tipuesearchWeights in the beginning, which was used in old algorithm.